refactor : Share Util 패키지 추가
