+++ core/common/src/main/kotlin/util/share/SharePreparer.kt
@@ -1,4 +1,4 @@
-package util
+package util.share
 
 import android.content.Context
 import android.graphics.Bitmap
